Safari-611.1.10/LayoutTests/media/video-poster-visible-after-first-video-frame-expected.txt


EXPECTED (internals.elementShouldDisplayPosterImage(video) == 'true') OK
EVENT(canplaythrough)
EXPECTED (internals.elementShouldDisplayPosterImage(video) == 'true') OK
EXPECTED (internals.elementShouldDisplayPosterImage(video) == 'false') OK
END OF TEST
